perf: unify DDP/single compile, add DDP optims, CNN channels_last auto-detect
P0: DDP compile path now uses compile_models() (backbone-only, reduce-overhead)
    instead of torch.compile(full_model, default mode), matching single-GPU behavior.

P1: DDP gains gradient_as_bucket_view=True (memory saving with find_unused=False)
    and prefetch_factor passthrough to DataLoader (config traceability).

P2: DPP gains static_graph=True (stable forward graph). DPObsEncoder auto-detects
    CNN backbones (ResNet/R3M) and applies channels_last memory format for ~15-22%
    conv speedup; ViT backbones (DINO/CLIP/SigLIP) are skipped.

Also includes: MoE logit overflow guard, diffusion v_pred float32 alphas cache,
PointNext hidden_channels floor fix, PointColorNoiseAug, StateNoiseAug clip_range,
MultiTask mp.Manager for spawn compatibility, assertion/NaN guard hardening.

All 6 strategies (dp/dp3/maniflow/moe_dp3/multitask_dit/r3d) smoke test PASSED.

+++ dexmani_policy/agents/action_decoders/backbone/ditx.py
@@ -396,7 +396,7 @@ class DiTXFlowMatch(nn.Module):
 
 
     def forward(self, x, timestep, target_t, context):
-        x = self.input_embedder(x) + self.input_pos_embed
+        x = self.input_embedder(x) + self.input_pos_embed.to(dtype=x.dtype)
 
         if not torch.is_tensor(timestep):
             timestep = torch.tensor([timestep], dtype=torch.float32, device=x.device)
@@ -414,7 +414,7 @@ class DiTXFlowMatch(nn.Module):
 
         time_c = self.timestep_and_target_t_fusion(torch.cat([timestep_embed, target_t_embed], dim=-1))
 
-        context_c = self.context_embedder(context) + self.context_pos_embed[:, :context.shape[1]]
+        context_c = self.context_embedder(context) + self.context_pos_embed[:, :context.shape[1]].to(dtype=context.dtype)
         if self.pre_norm_modality:
             context_c = self.context_norm(context_c, time_c)
         
@@ -542,7 +542,7 @@ class DiTXDiffusion(nn.Module):
 
 
     def forward(self, x, timestep, context):
-        x = self.input_embedder(x) + self.input_pos_embed
+        x = self.input_embedder(x) + self.input_pos_embed.to(dtype=x.dtype)
 
         if not torch.is_tensor(timestep):
             timestep = torch.tensor([timestep], dtype=torch.long, device=x.device)
@@ -553,7 +553,7 @@ class DiTXDiffusion(nn.Module):
 
         time_c = timestep_embed
 
-        context_c = self.context_embedder(context) + self.context_pos_embed[:, :context.shape[1]]
+        context_c = self.context_embedder(context) + self.context_pos_embed[:, :context.shape[1]].to(dtype=context.dtype)
         if self.pre_norm_modality:
             context_c = self.context_norm(context_c, time_c)
